--- a/litedram/gen-src/sdram_init/main.c
+++ b/litedram/gen-src/sdram_init/main.c
@@ -286,8 +286,7 @@ uint64_t main(void)
 	}
 	printf("\n");
 	if (ftr & SYS_REG_INFO_HAS_DRAM) {
-		printf("LiteDRAM built from Migen %s and LiteX %s\n",
-		       MIGEN_GIT_SHA1, LITEX_GIT_SHA1);
+		printf("LiteDRAM built from LiteX %s\n", LITEX_GIT_SHA1);
 		sdram_init();
 	}
 	if (ftr & SYS_REG_INFO_HAS_BRAM) {